requirements.txt
tensorflow
numpy
pillow
tensorflow-addons
gin-config